+++ packages/components-js/src/components/icon/icon/icon.tsx
@@ -1,10 +1,8 @@
 import { Build, Component, Element, h, Host, Prop, State, Watch } from '@stencil/core';
-import { getSvgContent, iconContent } from './icon-request';
+import { buildIconUrl, DEFAULT_ICON_NAME, getSvgContent } from './icon-request';
 import cx from 'classnames';
 import { prefix } from '../../../utils';
 import { Theme, IconName } from '../../../types';
-import { cdn, icons } from '@porsche-design-system/icons';
-import { camelCase } from 'change-case';
 
 @Component({
   tag: 'p-icon',
@@ -17,7 +15,7 @@ export class Icon {
   /**
    * Specifies which icon to use.
    */
-  @Prop() public name?: IconName = 'arrow-head-right';
+  @Prop() public name?: IconName = DEFAULT_ICON_NAME;
 
   /**
    * Specifies a whole icon path which can be used for custom icons.
@@ -47,10 +45,24 @@ export class Icon {
   /** Adapts the text color depending on the theme. Has no effect when "inherit" is set as color prop. */
   @Prop() public theme?: Theme = 'light';
 
-  private io?: IntersectionObserver;
-
   @State() private svgContent?: string;
   @State() private isVisible = false;
+
+  private io?: IntersectionObserver;
+
+  @Watch('source')
+  @Watch('name')
+  public loadIcon(): void {
+    if (Build.isBrowser && this.isVisible) {
+      this.svgContent = undefined; // reset svg content while new icon is loaded
+      const url = buildIconUrl(this.source ?? this.name);
+      getSvgContent(url).then((iconContent) => {
+        if (url === buildIconUrl(this.source ?? this.name)) { // check if response matches current icon source
+          this.svgContent = iconContent;
+        }
+      });
+    }
+  }
 
   public connectedCallback(): void {
     // purposely do not return the promise here because loading
@@ -69,36 +81,6 @@ export class Icon {
     }
   }
 
-  @Watch('source')
-  @Watch('name')
-  public loadIcon(): void {
-    if (Build.isBrowser && this.isVisible) {
-      const url = this.getSource();
-      if (iconContent.has(url)) {
-        // sync if it's already loaded
-        this.svgContent = iconContent.get(url);
-      } else {
-        // async if it hasn't been loaded
-        getSvgContent(url).then(() => {
-          if (url === this.getSource()) {
-            this.svgContent = iconContent.get(url);
-          }
-        });
-      }
-    }
-  }
-
-  public getSource(): string {
-    if (this.name && !this.source) {
-      return `${cdn}/${icons[camelCase(this.name)]}`;
-    }
-    if (this.source) {
-      return this.source;
-    }
-    console.warn('Please provide either an name property or a source property!');
-    return '';
-  }
-
   public render(): JSX.Element {
     const iconClasses = cx(
       prefix('icon'),
@@ -108,11 +90,8 @@ export class Icon {
     );
 
     return (
-      <Host role='img'>{(
-        (Build.isBrowser && this.svgContent)
-          ? <i class={iconClasses} innerHTML={this.svgContent}/>
-          : <i class={iconClasses}/>
-      )}
+      <Host role='img'>
+        <i class={iconClasses} innerHTML={this.svgContent}/>
       </Host>
     );
   }