+++ packages/components-js/tests/e2e/select-wrapper.e2e.ts
@@ -1,9 +1,14 @@
-import { E2EPage, newE2EPage } from '@stencil/core/testing';
+import {
+  getAttributeFromHandle,
+  getBoxShadow,
+  getClassFromHandle,
+  selectNode,
+  setContentWithDesignSystem, waitForInnerHTMLChange, waitForSelector
+} from './helpers';
 
 describe('select-wrapper', () => {
   it('should render', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper label="Some label">
         <select name="some-name">
           <option value="a">Option A</option>
@@ -12,13 +17,12 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>
     `);
-    const el = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
-    expect(el).not.toBeNull();
+    const el = await selectNode('p-select-wrapper >>> .p-select-wrapper__fake-select');
+    expect(el).toBeDefined();
   });
 
   it('should add aria-label to support screen readers properly', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper label="Some label">
         <select name="some-name">
           <option value="a">Option A</option>
@@ -27,13 +31,12 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>
     `);
-    const select = await page.find('p-select-wrapper select');
-    expect(select.getAttribute('aria-label')).toBe('Some label');
+    const select = await selectNode('p-select-wrapper select');
+    expect(await getAttributeFromHandle(select, 'aria-label')).toBe('Some label');
   });
 
   it('should add aria-label with description text to support screen readers properly', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper label="Some label" description="Some description">
         <select name="some-name">
           <option value="a">Option A</option>
@@ -42,13 +45,12 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>
     `);
-    const select = await page.find('p-select-wrapper select');
-    expect(select.getAttribute('aria-label')).toBe('Some label. Some description');
+    const select = await selectNode('p-select-wrapper select');
+    expect(await getAttributeFromHandle(select, 'aria-label')).toBe('Some label. Some description');
   });
 
   it('should add aria-label with message text to support screen readers properly', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper label="Some label" description="Some description" message="Some error message" state="error">
         <select name="some-name">
           <option value="a">Option A</option>
@@ -57,13 +59,12 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>
     `);
-    const select = await page.find('p-select-wrapper select');
-    expect(select.getAttribute('aria-label')).toBe('Some label. Some error message');
+    const select = await selectNode('p-select-wrapper select');
+    expect(await getAttributeFromHandle(select, 'aria-label')).toBe('Some label. Some error message');
   });
 
   it('should not render label if label prop is not defined but should render if changed programmatically', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper>
         <select name="some-name">
           <option value="a">Option A</option>
@@ -72,24 +73,17 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>`);
 
-    const selectComponent = await page.find('p-select-wrapper');
-    const getLabelText = async () => {
-      return selectComponent.shadowRoot.querySelector('.p-select-wrapper__label-text');
-    };
+    const selectComponent = await selectNode('p-select-wrapper');
+    const getLabelText = await selectNode('p-select-wrapper >>> .p-select-wrapper__label-text');
 
-    expect(await getLabelText()).toBeNull();
+    expect(getLabelText).toBeNull();
 
-    selectComponent.setProperty('label', 'Some label');
-
-    await page.waitForChanges();
-
-    expect(await getLabelText()).not.toBeNull();
-
+    await page.evaluate(el => el.setAttribute('label', 'Some label'), selectComponent);
+    expect(getLabelText).toBeDefined();
   });
 
   it('should add/remove message text and update aria-label attribute with message text if state changes programmatically', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`
+    await setContentWithDesignSystem(`
       <p-select-wrapper label="Some label">
         <select name="some-name">
           <option value="a">Option A</option>
@@ -98,48 +92,38 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>`);
 
-    const selectComponent = await page.find('p-select-wrapper');
-    const getMessage = async () => {
-      return selectComponent.shadowRoot.querySelector('.p-select-wrapper__message');
-    };
-
-    const getSelect = async () => {
-      return selectComponent.find('select');
-    };
+    const selectComponent = await selectNode('p-select-wrapper');
+    const getMessage = () => selectNode('p-select-wrapper >>> .p-select-wrapper__message'); // has to be a function because it only exists after first setAttribute
+    const select = await selectNode('select');
 
     expect(await getMessage()).toBeNull();
 
-    selectComponent.setProperty('state', 'error');
-    selectComponent.setProperty('message', 'Some error message');
+    await page.evaluate(el => el.setAttribute('state', 'error'), selectComponent);
+    await page.evaluate(el => el.setAttribute('message', 'Some error message'), selectComponent);
+    await waitForInnerHTMLChange(selectComponent);
 
-    await page.waitForChanges();
+    expect(await getMessage()).toBeDefined();
+    expect(await getAttributeFromHandle(await getMessage(), 'role')).toEqual('alert');
+    expect(await getAttributeFromHandle(select, 'aria-label')).toEqual('Some label. Some error message');
 
-    expect(await getMessage()).not.toBeNull();
-    expect(await getMessage()).toEqualAttribute('role', 'alert');
-    expect(await getSelect()).toEqualAttribute('aria-label','Some label. Some error message');
+    await page.evaluate(el => el.setAttribute('state', 'success'), selectComponent);
+    await page.evaluate(el => el.setAttribute('message', 'Some success message'), selectComponent);
+    await waitForInnerHTMLChange(selectComponent);
 
-    selectComponent.setProperty('state', 'success');
-    selectComponent.setProperty('message', 'Some success message');
+    expect(await getMessage()).toBeDefined();
+    expect(await getAttributeFromHandle(await getMessage(), 'role')).toBeNull();
+    expect(await getAttributeFromHandle(select, 'aria-label')).toEqual('Some label. Some success message');
 
-    await page.waitForChanges();
-
-    expect(await getMessage()).not.toBeNull();
-    expect(await getMessage()).not.toHaveAttribute('role');
-    expect(await getSelect()).toEqualAttribute('aria-label','Some label. Some success message');
-
-    selectComponent.setProperty('state', 'none');
-    selectComponent.setProperty('message', '');
-
-    await page.waitForChanges();
+    await page.evaluate(el => el.setAttribute('state', 'none'), selectComponent);
+    await page.evaluate(el => el.setAttribute('message', ''), selectComponent);
+    await waitForInnerHTMLChange(selectComponent);
 
     expect(await getMessage()).toBeNull();
-    expect(await getSelect()).toEqualAttribute('aria-label','Some label');
-
+    expect(await getAttributeFromHandle(select, 'aria-label')).toEqual('Some label');
   });
 
   it('should focus select when label text is clicked', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`<p-select-wrapper label="Some label">
+    await setContentWithDesignSystem(`<p-select-wrapper label="Some label">
       <select name="some-name">
         <option value="a">Option A</option>
         <option value="b">Option B</option>
@@ -147,8 +131,8 @@ describe('select-wrapper', () => {
       </select>
     </p-select-wrapper>`);
 
-    const labelText = await page.find('p-select-wrapper >>> .p-select-wrapper__label-text');
-    expect(labelText).not.toBeNull();
+    const labelText = await selectNode('p-select-wrapper >>> .p-select-wrapper__label-text');
+    expect(labelText).toBeDefined();
 
     async function hasSelectFocus() {
       return await page.evaluate(() => {
@@ -165,8 +149,7 @@ describe('select-wrapper', () => {
   });
 
   it('should disable fake select when select is set disabled programmatically', async () => {
-    const page = await newE2EPage();
-    await page.setContent(`<p-select-wrapper label="Some label">
+    await setContentWithDesignSystem(`<p-select-wrapper label="Some label">
       <select name="some-name">
         <option value="a">Option A</option>
         <option value="b">Option B</option>
@@ -174,47 +157,29 @@ describe('select-wrapper', () => {
       </select>
     </p-select-wrapper>`);
 
-    const getFakeSelect = async () => {
-      const selectWrapper = await page.find('p-select-wrapper');
-      return selectWrapper.shadowRoot.querySelector('.p-select-wrapper__fake-select');
-    };
-    const fakeSelectClassList = async () => (await getFakeSelect()).classList;
+    const fakeSelect = await selectNode('p-select-wrapper >>> .p-select-wrapper__fake-select');
+    const select = await selectNode('select');
 
-    expect((await getFakeSelect())).not.toHaveClass('p-select-wrapper__fake-select--disabled');
+    expect(await getClassFromHandle(fakeSelect)).not.toContain('p-select-wrapper__fake-select--disabled');
 
-    await page.evaluate(() => {
-      document.querySelector('select').disabled = true;
-    });
+    await select.evaluate((el: HTMLSelectElement) => el.disabled = true);
+    await waitForSelector(fakeSelect, 'p-select-wrapper__fake-select--disabled');
 
-    // for some reason we've to re-query the fakeSelect each time and .waitForSelector does not work
-    while(!(await fakeSelectClassList()).contains('p-select-wrapper__fake-select--disabled')) {
-      await page.waitFor(10);
-    }
+    expect(await getClassFromHandle(fakeSelect)).toContain('p-select-wrapper__fake-select--disabled');
 
-    expect((await getFakeSelect())).toHaveClass('p-select-wrapper__fake-select--disabled');
+    await select.evaluate((el: HTMLSelectElement) => el.disabled = false);
+    await waitForSelector(fakeSelect, 'p-select-wrapper__fake-select--disabled', {isGone: true});
 
-    await page.evaluate(() => {
-      document.querySelector('select').disabled = false;
-    });
-
-    // for some reason we've to re-query the fakeSelect each time and .waitForSelector does not work
-    while((await fakeSelectClassList()).contains('p-select-wrapper__fake-select--disabled')) {
-      await page.waitFor(10);
-    }
-
-    expect((await getFakeSelect())).not.toHaveClass('p-select-wrapper__fake-select--disabled');
+    expect(await getClassFromHandle(fakeSelect)).not.toContain('p-select-wrapper__fake-select--disabled');
   });
 
   describe('hover state', () => {
-    const getBoxShadow = async (page: E2EPage) => {
-      const fakeSelect = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
-      const styles = await fakeSelect.getComputedStyle();
-      return styles.boxShadow;
-    };
+
+    const getFakeSelect = () => selectNode('p-select-wrapper >>> .p-select-wrapper__fake-select');
 
     it('should change box-shadow color when fake select is hovered', async () => {
-      const page = await newE2EPage();
-      await page.setContent(`<p-select-wrapper label="Some label">
+      await page.reload();
+      await setContentWithDesignSystem(`<p-select-wrapper label="Some label">
         <select name="some-name">
           <option value="a">Option A</option>
           <option value="b">Option B</option>
@@ -222,18 +187,17 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>`);
 
-      const fakeSelect = await page.find('p-select-wrapper >>> .p-select-wrapper__fake-select');
-
-      const initialBoxShadow = await getBoxShadow(page);
+      const fakeSelect = await getFakeSelect();
+      const initialBoxShadow = await getBoxShadow(fakeSelect);
 
       await fakeSelect.hover();
 
-      expect(await getBoxShadow(page)).not.toBe(initialBoxShadow);
+      expect(await getBoxShadow(fakeSelect, {waitForTransition: true})).not.toBe(initialBoxShadow);
     });
 
     it('should change box-shadow color of fake select when label text is hovered', async () => {
-      const page = await newE2EPage();
-      await page.setContent(`<p-select-wrapper label="Some label">
+      await page.reload();
+      await setContentWithDesignSystem(`<p-select-wrapper label="Some label">
         <select name="some-name">
           <option value="a">Option A</option>
           <option value="b">Option B</option>
@@ -241,13 +205,13 @@ describe('select-wrapper', () => {
         </select>
       </p-select-wrapper>`);
 
-      const labelText = await page.find('p-select-wrapper >>> .p-select-wrapper__label-text');
-
-      const initialBoxShadow = await getBoxShadow(page);
+      const fakeSelect = await getFakeSelect();
+      const labelText = await selectNode('p-select-wrapper >>> .p-select-wrapper__label-text');
+      const initialBoxShadow = await getBoxShadow(fakeSelect);
 
       await labelText.hover();
 
-      expect(await getBoxShadow(page)).not.toBe(initialBoxShadow);
+      expect(await getBoxShadow(fakeSelect, {waitForTransition: true})).not.toBe(initialBoxShadow);
     });
   });
 });